requirements.txt
streamlit==1.37.0
ultralytics==8.2.0
opencv-python-headless
numpy
Pillow
torch
torchvision